// bin/modules/smartedit/smartedit/apps/smartedit-commons/src/components/genericEditor/types.ts
export interface GenericEditorAttribute {
    cmsStructureType: string;
    cmsStructureEnumType?: string;
    qualifier: string;
    i18nKey?: string;
    localized?: boolean;
    editable?: boolean;
    required?: boolean;
    collection?: boolean;
    postfixText?: string;
}

export enum GenericEditorMediaType {
    VIDEO = 'video',
    IMAGE = 'image',
    PDF_DOCUMENT = 'pdf',
    DEFAULT = ''
}

/**
 * **Deprecated since 2105. It will be removed in next release.**
 *
 * custom sanitization function
 * first argument is a property value (not the full model) that may be comple or scalar
 *
 * @deprecated
 */
export type GenericEditorCustomSanitize<T = any> = (
    payload: T,
    sanitize: (input: string) => string
) => T;

/** Describes the field being edited as defined by the structure API described in {@link GenericEditorFactoryService}. */
export interface GenericEditorField extends GenericEditorAttribute {
    editorStackId?: string;
    hasErrors?: boolean;
    hasWarnings?: boolean;
    smarteditComponentType?: string;
    messages?: GenericEditorFieldMessage[];
    /**
     * A boolean to determine if we are in paged mode as opposed to retrieving all items at once.
     */
    paged?: boolean;
    template?: string;
    component?: Type<any>;
    initiated?: string[];
    /** The original array of options (used by {@link OptionsDropdownPopulator}). */
    options?: TypedMap<string[]> | GenericEditorOption;
    hidePrefixLabel?: boolean;
    hideFieldWidget?: boolean;
    dependsOnValue?: string;
    dependsOnField?: string;
    allowMediaType?: GenericEditorMediaType;
    /**
     * **Deprecated since 2105. It will be removed in next release.**
     *
     * @deprecated
     */
    customSanitize?: GenericEditorCustomSanitize;
    requiresUserCheck?: TypedMap<boolean>;
    isUserChecked?: boolean;
    defaultValue?: string;
    params?: TypedMap<string>;
    /**
     * A comma separated list of attributes to include from the model when building the request params.
     * The qualifier of the parent dropdown that this dropdown depends on.
     */
    dependsOn?: string;
    /**
     * The uri used to make a rest call to fetch data (used by {@link UriDropdownPopulator})
     */
    uri?: string;
    /**
     * If a propertyType is defined, the `GenericEditorDropdownComponent` will use the populator associated to it with the following recipe name: `propertyType + "DropdownPopulator"`.
     *
     * For AngularJS `SeDropdownComponent`.
     */
    propertyType?: string;
    /**
     * The name of the attribute to use when setting the id attribute.
     */
    idAttribute?: string;
    /**
     * A list of attributes to use when setting the label attribute.
     */
    labelAttributes?: string[];
    errors?: GenericEditorFieldMessage[];
    control?: FormControl;
    validators?: GenericEditorFieldValidatorMap;

    /**
     * This map is only used for localized fields. Each entry contains a boolean
     * that specifies whether the field should be enabled for a language or not.
     */
    isLanguageEnabledMap?: { [languageId: string]: boolean };
    tooltip?: string;
    labelText?: string;
    placeholder?: string;
    containedTypes?: string[];
}
// ...
export interface GenericEditorFieldMessage {
    fromSubmit?: boolean;
    isNonPristine?: boolean;
    message: string;
    subject?: string;
    type?: string;
    uniqId?: string;
    marker?: string;
    format?: string;
    language?: string;
    [index: string]: any;
}
// ...
export type GenericEditorFieldValidatorPredicate = (
    id: string,
    structure: GenericEditorField,
    required: boolean,
    component: Payload
) => boolean;

export type GenericEditorFieldValidatorMap = TypedMap<GenericEditorFieldValidatorPredicate>;
// ...
export interface GenericEditorOption {
    id?: string;
    label?: string | TypedMap<string>;
    [key: string]: any;
}